puse que el cliente por defecto sea consumidor final en ventas, con f1 entre a ventas, location.reload en ventasy y en agregar cliente y modifiar un select

+++ src/productos/productos.js
@@ -90,9 +90,11 @@ ipcRenderer.on('informacion',(e,args)=>{
 
 eliminar.addEventListener('click',async e=>{
     if(productoSeleccionado){
-        const mensaje = (await axios.delete(`${URL}productos/${productoSeleccionado.id}`)).data;
-        alert(mensaje);
-        location.reload();
+        if (confirm("Seguro Borrar Producto")) {
+            const mensaje = (await axios.delete(`${URL}productos/${productoSeleccionado.id}`)).data;
+            alert(mensaje);
+            location.reload();
+        }
     }else{
         alert("No se selecciono ningun producto");
     }
